fixed total calculation

+++ software/database.sql
@@ -4,7 +4,7 @@ USE finance_data;
 
 -- Create the users table if it does not exist
 CREATE TABLE IF NOT EXISTS user (
-    user_id INT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT AUTO_INCREMENT PRIMARY KEY,
     username VARCHAR(50) NOT NULL,
     email VARCHAR(100) NOT NULL
 );
